agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Restaurant Finder Agent with A2UI support"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "a2ui-agent",
    "ag-ui-adk>=0.4.2",
    "fastapi>=0.123.10",
    "google-adk>=1.23.0",
    "jsonschema>=4.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.14.14",
]

[tool.uv.sources]
a2ui-agent = { git = "https://github.com/google/A2UI.git", subdirectory = "a2a_agents/python/a2ui_agent" }
